feat: modularize audit batching, observability, and telemetry endpoints
- Refactored BankAccountServiceImpl to delegate queue management to AuditProcessingService
- Injected MetricsCollector for observability at transaction ingestion and audit submission
- Decoupled batching logic using pluggable BatchingAlgorithm with bin-packing optimization
- Introduced AuditBatch model with batchId and value metadata
- Updated AuditProcessingService to use bounded queue and invoke external persistence + notifier
- Implemented ConsoleAuditNotifier to emit structured audit logs and trigger telemetry events
- Added AuditStatsService for tracking batch history, submission count, and telemetry records
- Exposed REST endpoints for transaction history, balance view, audit batches, and telemetry
- Applied DDD-aligned packaging: account, audit, shared.observability
- Added atomic operations and thread-safe structures (AtomicReference, CopyOnWriteArrayList)
- Future-proofed architecture for Kafka migration and microservice isolation

BREAKING CHANGE:
BankAccountServiceImpl no longer manages audit lifecycle threads;
audit submission now orchestrated externally via AuditProcessingService

+++ src/main/java/com/jpc/exercise/audit/domain/model/AuditBatch.java
@@ -21,9 +21,9 @@ public class AuditBatch {
         return transactions;
     }
 
-    public long getTotalValue() {
+    public double getTotalValue() {
         return transactions.stream()
-                .mapToLong(tx -> Math.abs(tx.getAmount()))
+                .mapToDouble(tx -> Math.abs(tx.getAmount()))
                 .sum();
     }
